Rename install_date => insttime
